changed all instance of tables into tabular
