+++ proto-google-cloud-aiplatform-v1beta1/src/main/java/com/google/cloud/aiplatform/v1beta1/BatchPredictionJob.java
@@ -2016,10 +2016,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
      * BigQuery-dataset-name compatible (for example, most special characters
      * become underscores), and timestamp is in
      * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-     * two tables will be created, `predictions`, and `errors`.
+     * two tabular will be created, `predictions`, and `errors`.
      * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
      * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-     * defined then the tables have columns as follows: The `predictions`
+     * defined then the tabular have columns as follows: The `predictions`
      * table contains instances for which the prediction succeeded, it
      * has columns as per a concatenation of the Model's instance and
      * prediction schemata. The `errors` table contains rows for which the
@@ -2045,10 +2045,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
      * BigQuery-dataset-name compatible (for example, most special characters
      * become underscores), and timestamp is in
      * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-     * two tables will be created, `predictions`, and `errors`.
+     * two tabular will be created, `predictions`, and `errors`.
      * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
      * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-     * defined then the tables have columns as follows: The `predictions`
+     * defined then the tabular have columns as follows: The `predictions`
      * table contains instances for which the prediction succeeded, it
      * has columns as per a concatenation of the Model's instance and
      * prediction schemata. The `errors` table contains rows for which the
@@ -2074,10 +2074,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
      * BigQuery-dataset-name compatible (for example, most special characters
      * become underscores), and timestamp is in
      * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-     * two tables will be created, `predictions`, and `errors`.
+     * two tabular will be created, `predictions`, and `errors`.
      * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
      * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-     * defined then the tables have columns as follows: The `predictions`
+     * defined then the tabular have columns as follows: The `predictions`
      * table contains instances for which the prediction succeeded, it
      * has columns as per a concatenation of the Model's instance and
      * prediction schemata. The `errors` table contains rows for which the
@@ -2430,10 +2430,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
      * BigQuery-dataset-name compatible (for example, most special characters
      * become underscores), and timestamp is in
      * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-     * two tables will be created, `predictions`, and `errors`.
+     * two tabular will be created, `predictions`, and `errors`.
      * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
      * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-     * defined then the tables have columns as follows: The `predictions`
+     * defined then the tabular have columns as follows: The `predictions`
      * table contains instances for which the prediction succeeded, it
      * has columns as per a concatenation of the Model's instance and
      * prediction schemata. The `errors` table contains rows for which the
@@ -2462,10 +2462,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
      * BigQuery-dataset-name compatible (for example, most special characters
      * become underscores), and timestamp is in
      * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-     * two tables will be created, `predictions`, and `errors`.
+     * two tabular will be created, `predictions`, and `errors`.
      * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
      * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-     * defined then the tables have columns as follows: The `predictions`
+     * defined then the tabular have columns as follows: The `predictions`
      * table contains instances for which the prediction succeeded, it
      * has columns as per a concatenation of the Model's instance and
      * prediction schemata. The `errors` table contains rows for which the
@@ -2497,10 +2497,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
      * BigQuery-dataset-name compatible (for example, most special characters
      * become underscores), and timestamp is in
      * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-     * two tables will be created, `predictions`, and `errors`.
+     * two tabular will be created, `predictions`, and `errors`.
      * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
      * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-     * defined then the tables have columns as follows: The `predictions`
+     * defined then the tabular have columns as follows: The `predictions`
      * table contains instances for which the prediction succeeded, it
      * has columns as per a concatenation of the Model's instance and
      * prediction schemata. The `errors` table contains rows for which the
@@ -3401,10 +3401,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3433,10 +3433,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3475,10 +3475,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3515,10 +3515,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3552,10 +3552,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3602,10 +3602,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3644,10 +3644,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3674,10 +3674,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
@@ -3712,10 +3712,10 @@ public final class BatchPredictionJob extends com.google.protobuf.GeneratedMessa
        * BigQuery-dataset-name compatible (for example, most special characters
        * become underscores), and timestamp is in
        * YYYY_MM_DDThh_mm_ss_sssZ "based on ISO-8601" format. In the dataset
-       * two tables will be created, `predictions`, and `errors`.
+       * two tabular will be created, `predictions`, and `errors`.
        * If the Model has both [instance][google.cloud.aiplatform.v1beta1.PredictSchemata.instance_schema_uri]
        * and [prediction][google.cloud.aiplatform.v1beta1.PredictSchemata.parameters_schema_uri] schemata
-       * defined then the tables have columns as follows: The `predictions`
+       * defined then the tabular have columns as follows: The `predictions`
        * table contains instances for which the prediction succeeded, it
        * has columns as per a concatenation of the Model's instance and
        * prediction schemata. The `errors` table contains rows for which the
