+++ proto-google-cloud-aiplatform-v1beta1/src/main/java/com/google/cloud/aiplatform/v1beta1/DeployedModelOrBuilder.java
@@ -285,7 +285,7 @@ public interface DeployedModelOrBuilder
    * <pre>
    * If true, the container of the DeployedModel instances will send `stderr`
    * and `stdout` streams to Stackdriver Logging.
-   * Only supported for custom-trained Models and AutoML Tables Models.
+   * Only supported for custom-trained Models and AutoML Tabular Models.
    * </pre>
    *
    * <code>bool enable_container_logging = 12;</code>
