+++ proto-google-cloud-aiplatform-v1beta1/src/main/proto/google/cloud/aiplatform/v1beta1/endpoint.proto
@@ -134,7 +134,7 @@ message DeployedModel {
   // If true, the container of the DeployedModel instances will send `stderr`
   // and `stdout` streams to Stackdriver Logging.
   //
-  // Only supported for custom-trained Models and AutoML Tables Models.
+  // Only supported for custom-trained Models and AutoML Tabular Models.
   bool enable_container_logging = 12;
 
   // These logs are like standard server access logs, containing
