+++ samples/snippets/src/main/java/aiplatform/CreateDatasetTabularGcsSample.java
@@ -16,7 +16,7 @@
 
 package aiplatform;
 
-// [START aiplatform_create_dataset_tables_gcs_sample]
+// [START aiplatform_create_dataset_tabular_gcs_sample]
 
 import com.google.api.gax.longrunning.OperationFuture;
 import com.google.cloud.aiplatform.v1beta1.CreateDatasetOperationMetadata;
@@ -31,7 +31,7 @@ import java.util.concurrent.ExecutionException;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.TimeoutException;
 
-public class CreateDatasetTablesGcsSample {
+public class CreateDatasetTabularGcsSample {
 
   public static void main(String[] args)
       throws InterruptedException, ExecutionException, TimeoutException, IOException {
@@ -85,4 +85,4 @@ public class CreateDatasetTablesGcsSample {
     }
   }
 }
-// [END aiplatform_create_dataset_tables_gcs_sample]
+// [END aiplatform_create_dataset_tabular_gcs_sample]
