+++ samples/snippets/src/main/java/aiplatform/CreateTrainingPipelineTabularClassificationSample.java
@@ -16,7 +16,7 @@
 
 package aiplatform;
 
-// [START aiplatform_create_training_pipeline_tables_classification_sample]
+// [START aiplatform_create_training_pipeline_tabular_classification_sample]
 
 import com.google.cloud.aiplatform.v1beta1.DeployedModelRef;
 import com.google.cloud.aiplatform.v1beta1.EnvVar;
@@ -44,7 +44,7 @@ import com.google.rpc.Status;
 import java.io.IOException;
 import java.util.List;
 
-public class CreateTrainingPipelineTablesClassificationSample {
+public class CreateTrainingPipelineTabularClassificationSample {
 
   public static void main(String[] args) throws IOException {
     // TODO(developer): Replace these variables before running the sample.
@@ -115,7 +115,7 @@ public class CreateTrainingPipelineTablesClassificationSample {
       TrainingPipeline trainingPipelineResponse =
           pipelineServiceClient.createTrainingPipeline(locationName, trainingPipeline);
 
-      System.out.println("Create Training Pipeline Tables Classification Response");
+      System.out.println("Create Training Pipeline Tabular Classification Response");
       System.out.format("\tName: %s\n", trainingPipelineResponse.getName());
       System.out.format("\tDisplay Name: %s\n", trainingPipelineResponse.getDisplayName());
       System.out.format(
@@ -252,4 +252,4 @@ public class CreateTrainingPipelineTablesClassificationSample {
     }
   }
 }
-// [END aiplatform_create_training_pipeline_tables_classification_sample]
+// [END aiplatform_create_training_pipeline_tabular_classification_sample]
