+++ samples/snippets/src/test/java/aiplatform/CreateTrainingPipelineTabularRegressionSampleTest.java
@@ -31,29 +31,27 @@ import org.junit.Before;
 import org.junit.BeforeClass;
 import org.junit.Test;
 
-public class CreateTrainingPipelineTablesClassificationSampleTest {
+public class CreateTrainingPipelineTabularRegressionSampleTest {
 
   private static final String PROJECT = System.getenv("UCAIP_PROJECT_ID");
   private static final String DATASET_ID =
-      System.getenv("TRAINING_PIPELINE_TABLES_CLASSIFICATION_DATASET_ID");
-  private static final String TARGET_COLUMN = "TripType";
+          System.getenv("TRAINING_PIPELINE_TABLES_REGRESSION_DATASET_ID");
+  private static final String TARGET_COLUMN = "Amount";
   private static final String TRANSFORMATION =
-      "[{\"numeric\":{\"columnName\":\"Age\",\"invalidValuesAllowed\":false}},"
-              + "{\"categorical\":{\"columnName\":\"Job\"}},"
-              + "{\"categorical\":{\"columnName\":\"MaritalStatus\"}},"
-              + "{\"categorical\":{\"columnName\":\"Default\"}},"
-              + "{\"numeric\":{\"columnName\":\"Balance\",\"invalidValuesAllowed\":false}},"
-              + "{\"categorical\":{\"columnName\":\"Housing\"}},"
-              + "{\"categorical\":{\"columnName\":\"Loan\"}},"
-              + "{\"categorical\":{\"columnName\":\"Contact\"}},"
-              + "{\"numeric\":{\"columnName\":\"Day\",\"invalidValuesAllowed\":false}},"
-              + "{\"categorical\":{\"columnName\":\"Month\"}},"
-              + "{\"numeric\":{\"columnName\":\"Duration\",\"invalidValuesAllowed\":false}},"
-              + "{\"numeric\":{\"columnName\":\"Campaign\",\"invalidValuesAllowed\":false}},"
-              + "{\"numeric\":{\"columnName\":\"PDays\",\"invalidValuesAllowed\":false}},"
-              + "{\"numeric\":{\"columnName\":\"Previous\",\"invalidValuesAllowed\":false}},"
-              + "{\"categorical\":{\"columnName\":\"POutcome\"}},"
-              + "{\"categorical\":{\"columnName\":\"Deposit\"}}]";
+      "[{\"categorical\":{\"columnName\":\"SC_Group_Desc\"}},"
+              + "{\"categorical\":{\"columnName\":\"SC_GroupCommod_ID\"}},"
+              + "{\"categorical\":{\"columnName\":\"SC_GroupCommod_Desc\"}},"
+              + "{\"numeric\":{\"columnName\":\"SortOrder\",\"invalidValuesAllowed\":false}},"
+              + "{\"text\":{\"columnName\":\"SC_GeographyIndented_Desc\"}},"
+              + "{\"numeric\":{\"columnName\":\"SC_Commodity_ID\",\"invalidValuesAllowed\":false}},"
+              + "{\"text\":{\"columnName\":\"SC_Commodity_Desc\"}},"
+              + "{\"numeric\":{\"columnName\":\"SC_Attribute_ID\",\"invalidValuesAllowed\":false}},"
+              + "{\"text\":{\"columnName\":\"SC_Attribute_Desc\"}},"
+              + "{\"numeric\":{\"columnName\":\"SC_Unit_ID\",\"invalidValuesAllowed\":false}},"
+              + "{\"numeric\":{\"columnName\":\"Year_ID\",\"invalidValuesAllowed\":false}},"
+              + "{\"categorical\":{\"columnName\":\"SC_Frequency_Desc\"}},"
+              + "{\"numeric\":{\"columnName\":\"Timeperiod_ID\",\"invalidValuesAllowed\":false}},"
+              + "{\"text\":{\"columnName\":\"Timeperiod_Desc\"}}]";
   private ByteArrayOutputStream bout;
   private PrintStream out;
   private PrintStream originalPrintStream;
@@ -69,7 +67,7 @@ public class CreateTrainingPipelineTablesClassificationSampleTest {
   public static void checkRequirements() {
     requireEnvVar("GOOGLE_APPLICATION_CREDENTIALS");
     requireEnvVar("UCAIP_PROJECT_ID");
-    requireEnvVar("TRAINING_PIPELINE_TABLES_CLASSIFICATION_DATASET_ID");
+    requireEnvVar("TRAINING_PIPELINE_TABLES_REGRESSION_DATASET_ID");
   }
 
   @Before
@@ -102,20 +100,20 @@ public class CreateTrainingPipelineTablesClassificationSampleTest {
   }
 
   @Test
-  public void createTrainingPipelineTablesClassification() throws IOException {
+  public void createTrainingPipelineTabularRegression() throws IOException {
     // Act
     String modelDisplayName =
         String.format(
-            "temp_create_training_pipeline_tables_classification_test_%s",
+            "temp_create_training_pipelinetabularregression_model_test_%s",
             UUID.randomUUID().toString().replaceAll("-", "_").substring(0, 26));
 
-    CreateTrainingPipelineTablesClassificationSample.createTrainingPipelineTableClassification(
+    CreateTrainingPipelineTabularRegressionSample.createTrainingPipelineTableRegression(
         PROJECT, modelDisplayName, DATASET_ID, TARGET_COLUMN, TRANSFORMATION);
 
     // Assert
     String got = bout.toString();
     assertThat(got).contains(DATASET_ID);
-    assertThat(got).contains("Create Training Pipeline Tables Classification Response");
+    assertThat(got).contains("Create Training Pipeline Tabular Regression Response");
     trainingPipelineId = got.split("Name: ")[1].split("trainingPipelines/")[1].split("\n")[0];
   }
 }
